requirements.txt
# FastAPI 核心框架
fastapi==0.104.1
uvicorn[standard]==0.24.0

# 数据库相关
sqlalchemy==2.0.23
alembic==1.12.1

# 数据验证和序列化
pydantic==2.5.0
pydantic-settings==2.1.0

# HTTP 客户端
httpx==0.25.2
requests==2.31.0

# 异步任务队列
celery==5.3.4
redis==5.0.1

# 环境变量管理
python-dotenv==1.0.0

# AI 模型集成
openai==1.3.7
google-generativeai==0.3.2

# JWT 认证
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# 图像处理和OCR
Pillow==10.1.0
pytesseract==0.3.10
opencv-python==4.8.1.78

# 文件处理
python-magic==0.4.27
PyPDF2==3.0.1
python-docx==1.1.0

# WebSocket支持
websockets==12.0

# 日期时间处理
python-dateutil==2.8.2

# 数据处理
pandas==2.1.4
numpy==1.25.2
openpyxl==3.1.5

# 日志记录
loguru==0.7.2

# 测试框架
pytest==7.4.3
pytest-asyncio==0.21.1

# 开发工具
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# 监控和性能
prometheus-client==0.19.0

# 邮件发送
fastapi-mail==1.4.1

# 缓存
cachetools==5.3.2

# 配置管理
pyyaml==6.0.1

# 安全相关
cryptography==41.0.7

# 时区处理
pytz==2023.3

# 数据导出
openpyxl==3.1.2

# API文档增强
fastapi-users==12.1.2

# 限流
slowapi==0.1.9

# 健康检查
fastapi-health==0.4.0